air-pollution-analysis-tool/satellite-data-core/requirements.txt
aiofiles==0.6.0
aioify==0.4.0
aniso8601==9.0.1
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1602546592129/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1599308529326/work
awaits==0.0.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache==1.6.1
basemap @ file:///home/conda/feedstock_root/build_artifacts/basemap_1602682525274/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
blinker==1.4
bokeh @ file:///home/conda/feedstock_root/build_artifacts/bokeh_1603280822353/work
brotlipy==0.7.0
Cartopy @ file:///home/conda/feedstock_root/build_artifacts/cartopy_1603098797244/work
certifi==2020.6.20
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1602537222527/work
cftime @ file:///home/conda/feedstock_root/build_artifacts/cftime_1602504445174/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1602255302199/work
click==7.1.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1602614066034/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1602029610262/work
decorator==4.4.2
defusedxml==0.6.0
distributed @ file:///home/conda/feedstock_root/build_artifacts/distributed_1602493192806/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1602701739019/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask==1.1.2
Flask-RESTful==0.3.8
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1602700749102/work
future==0.18.2
GDAL==3.1.3
geojson==2.5.0
geomet @ file:///home/conda/feedstock_root/build_artifacts/geomet_1601766976124/work
h11==0.12.0
h2==4.0.0
HeapDict==1.0.1
hpack==4.0.0
html2text==2020.1.16
Hypercorn==0.11.2
hyperframe==6.0.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1602263269022/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1602682806089/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1602640391633/work
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1602395232480/work
Jinja2==2.11.2
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1598486169312/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1602537278122/work
jupyterlab==2.2.8
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1593951277307/work
kafka-python==2.0.2
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1603882994700/work
locket==0.2.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1602267312178/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1602600751051/work
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1602381748886/work
modapsclient==0.1.1
module-wrapper==0.3.1
msgpack @ file:///home/conda/feedstock_root/build_artifacts/msgpack-python_1602380756791/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1602859080374/work
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert_1602715394210/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1602732862338/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1601342677072/work
netCDF4 @ file:///home/conda/feedstock_root/build_artifacts/netcdf4_1602508541961/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1602720145052/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1588605082776/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
panda==0.3.1
pandas==1.1.3
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1595548966091/work
partd==1.1.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1603406598992/work
priority==1.3.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1602524994744/work
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1603570988815/work
ptyprocess==0.6.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1600347314331/work
pyhdf @ file:///home/conda/feedstock_root/build_artifacts/pyhdf_1602545641568/work
pyModis==2.1.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj @ file:///home/conda/feedstock_root/build_artifacts/pyproj_1602664229713/work
pyrsistent @ file:///home/conda/feedstock_root/build_artifacts/pyrsistent_1602259974442/work
pyshp @ file:///home/conda/feedstock_root/build_artifacts/pyshp_1599782465740/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1602326928339/work
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
Quart==0.14.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
Send2Trash==1.5.0
sentinelsat @ file:///home/conda/feedstock_root/build_artifacts/sentinelsat_1601821879944/work
Shapely @ file:///home/conda/feedstock_root/build_artifacts/shapely_1602547954120/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1591999956871/work
stdlib-list==0.8.0
tblib==1.6.0
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1602679586280/work
testpath==0.4.4
toml==0.10.2
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1600973991856/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1602488891005/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1603658198734/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1603125704209/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
wsproto==1.0.0
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1600638299066/work
zict==2.0.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1602852756910/work
